Moved single node specific entities in server::partition to nested module.
