+++ src/server/partition/mod.rs
@@ -27,18 +27,4 @@ pub struct PartitionId {
     pub partition_number: u64,
 }
 
-pub mod single_node {
-    use std::{borrow::Cow, time::Duration};
-
-    pub enum PartitionRequest {
-        RemoveExpired { expiry_duration: Duration },
-
-        Read { offset: u64 },
-        Append { record_bytes: Cow<'static, [u8]> },
-
-        LowestOffset,
-        HighestOffset,
-    }
-}
-
-pub mod in_memory;
+pub mod single_node;
